+++ src/main/java/com/kakaotech/team14backend/image/infrastructure/ImageRepository.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.image.repository;
+package com.kakaotech.team14backend.image.infrastructure;
 
 import com.kakaotech.team14backend.image.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
